Fix date parsing
